refactor(analytics-control-bar): extract date picker and interval filters into their own files
