+++ src/components/analytics-control-bar-date-range-filter/README.md
@@ -1,0 +1,4 @@
+# AnalyticsControlBarDateRangeFilter
+
+A filter displayed in the analytics control bar that allows for relative and absolute date range
+selection.
