+++ src/components/analytics-control-bar-date-range-filter/styles.module.scss
@@ -1,0 +1,40 @@
+@import "@density/ui/variables/colors.scss";
+
+.header {
+  width: 400px;
+  background-color: $grayLightest;
+}
+.back {
+  margin-right: 16px;
+}
+
+.dateSelector {
+  white-space: nowrap;
+}
+.dateSelectorLabel {
+  color: $gray;
+  padding-left: 16px;
+  padding-right: 16px;
+}
+
+.dateSelectorCustomWrapper {
+  display: flex;
+  flex-direction: column;
+  height: 564px;
+  width: 400px;
+}
+.dateSelectorCustomBody {
+  flex-grow: 1;
+  flex-shrink: 1;
+
+  display: flex;
+  flex-direction: column;
+  padding: 24px;
+}
+.dateSelectorCustomBodyItem {
+  padding-top: 4px;
+  padding-bottom: 4px;
+  flex-grow: 0;
+  flex-shrink: 0;
+}
+.dateSelectorCustomBodyItem:first-child { padding-top: 0px; }
