+++ src/components/analytics-control-bar-interval-filter/README.md
@@ -1,0 +1,4 @@
+# AnalyticsControlBarIntervalFilter
+
+Defines a simple filter in the analytics control bar that renders a list of intervals to be used
+when querying count data.
