+++ src/components/analytics-control-bar-interval-filter/index.tsx
@@ -2,8 +2,7 @@ import React, { useState } from 'react';
 import styles from './styles.module.scss';
 
 import Filter, { FilterBold } from '../analytics-control-bar-filter';
-import { ItemList } from './utilities';
-
+import { ItemList } from '../analytics-control-bar/utilities';
 
 export enum AnalyticsInterval {
   DAY = 'DAY',
